==> Notebooks/03_deconvulation.Rmd <==
Jai Babe Di
Jai Guru Maa Ji
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(SpatialDecon)
library(GeomxTools)
library(reshape2)
#https://bioconductor.org/packages/devel/bioc/vignettes/SpatialDecon/inst/doc/SpatialDecon_vignette_NSCLC.html
```

```{r}
#A “cell profile matrix” is a pre-defined matrix that specifies the expected expression profiles of each cell type in the experiment. The SpatialDecon library comes with one such matrix pre-loaded, the “SafeTME” matrix, designed for estimation of immune and stroma cells in the tumor microenvironment.
data("safeTME")
data("safeTME.matches")
```

```{r}

target_demoData_OS_sample <- readRDS("~/Desktop/GeoMX_TMA/output_files/revision_data/01_normalized_OS_samples_filtered.RDS")

```


```{r}

target_demoData_OS_sample_lowDFI <- subset(target_demoData_OS_sample, select = sData(target_demoData_OS_sample)[["DFI_sum"]] %in% "low_DFI") 

target_demoData_OS_sample_highDFI <- subset(target_demoData_OS_sample, select = sData(target_demoData_OS_sample)[["DFI_sum"]] %in% "high_DFI") 

sample_run_list <- c("target_demoData_OS_sample_lowDFI", "target_demoData_OS_sample_highDFI")

GOI_heatmap <- lapply(sample_run_list, function(sample_name){
  
sample_name <- "target_demoData_OS_sample_highDFI"
sample_object <- get(sample_name)

# run spatialdecon with all the bells and whistles:
restils = runspatialdecon(object = sample_object,
                          norm_elt = "q_norm",  # normalized data
                          raw_elt = "exprs",    # expected background counts for every data point in norm
                          X = safeTME,                            # safeTME matrix, used by default
                          cellmerges = safeTME.matches,           # safeTME.matches object, used by default
                          cell_counts = sample_object[["Nuclei"]], # nuclei counts, used to estimate total cells
                          n_tumor_clusters = 5) 

#res@phenoData@data$prop_of_nontumor
# change sample name to roi
sample_object@phenoData@data$`Scan Name` <- gsub("Osteosarcoma TMA ","",sData(sample_object)[,"Scan Name"])
sample_object@protocolData@data$sample_id <- paste(sample_object@protocolData@data$Roi,sample_object@phenoData@data$`Scan Name`, sep = ".")

df <- data.frame("file_name"= rownames(sData(sample_object)),
"Roi"=sData(sample_object)[,"sample_id"])

i1 <- match(rownames(restils$prop_of_all), df$file_name)
i2 <- !is.na(i1) # to take care of non matches which are NA
rownames(restils$prop_of_all)[i2] <- df$Roi[i1[i2]]

# or the proportions of cells:
temp = replace(restils$prop_of_all, is.na(restils$prop_of_all), 0)
o = hclust(dist(temp))$order

print(sample_name)

data_long <- melt(restils$prop_of_all, id.vars = "type", 
                  variable.name = "Cell_Type", 
                  value.name = "Value")
data_long$type <- gsub("target_demoData_OS_sample_","",sample_name)
data_long$Value <- as.numeric(data_long$Value)
data_long$Var2 <- as.factor(data_long$Var2)

write.csv(data_long,paste0("~/Desktop/GeoMX_TMA/output_files/revision_data/",sample_name,".csv"), quote = F,row.names = F)

pdf(paste0("~/Desktop/GeoMX_TMA/plots/revision_data/deconvulation_barplot_check",sample_name,".pdf")) 
TIL_barplot(t(restils$prop_of_all)[, o],
            draw_legend = TRUE, cex.names = 0.6)


write.csv(restils$prop_of_all,paste0("~/Desktop/GeoMX_TMA/output_files/revision_data/deconvulation_output_file",sample_name,".csv"), quote = F)
})
```

```{r}

library(ggpubr)
high_dfi_df <- read.csv("~/Desktop/GeoMX_TMA/output_files/revision_data/target_demoData_OS_sample_highDFI.csv", sep = ",")
low_dfi_df <- read.csv("~/Desktop/GeoMX_TMA/output_files/revision_data/target_demoData_OS_sample_lowDFI.csv", sep = ",")

both_dfi <- rbind(high_dfi_df, low_dfi_df)
  
# Create boxplots with p-values
p <- ggplot(both_dfi, aes(x = type, y = Value, fill = type)) +
  geom_boxplot(outlier.shape = NA, alpha = 0.7) + 
  geom_jitter(width = 0.2, size = 0.5, alpha = 0.6) + 
  facet_wrap(~ Var2, scales = "free_y") +
  labs(title = "Comparison of Cell Types Between low and high DFI",
       x = "Type", 
       y = "Proportions") +
  theme_minimal(base_size = 14) +
  
  scale_fill_manual(values = c("lowDFI" = "steelblue", "highDFI" = "tomato")) +
  theme(axis.title = element_text(face = "bold"),
        plot.title = element_text(size = 12, face = "bold"),
        axis.text.x=element_blank())+
 stat_compare_means(aes(label = format(p.format, scientific = TRUE)), method = "wilcox.test",label = "p.format", 
    label.sep = " = ") 

#pdf("~/Desktop/GeoMX_TMA/plots/revision_data/high_low_dfi_deconv_barplots.pdf", bg = "white", height = 12 , width = 9)
print(p)
```
